requirements.txt
annotated-types==0.6.0
anyio==4.0.0
asgiref==3.7.2
astroid==2.15.8
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
colorama==0.4.6
copier==8.3.0
cryptography==41.0.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
Django==4.2.4
django-filter==23.3
django-oauth-toolkit==2.3.0
django-taggit==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
dunamai==1.19.0
ecdsa==0.18.0
flet==0.10.3
flet-core==0.10.3
flet-runtime==0.10.3
funcy==2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
isort==5.12.0
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
jwcrypto==1.5.0
jwt==1.3.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
plumbum==1.8.2
prompt-toolkit==3.0.36
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pypng==0.20220715.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.1
repath==0.9.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomlkit==0.12.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.4
watchdog==3.0.0
wcwidth==0.2.8
websocket-client==1.6.4
websockets==11.0.3
wrapt==1.15.0
